move von_neumann_entropy to qis.measures
